Use 32-bit format for CTM
Using 32-bits isn't the best, but it better represents the S31.32
format than 16-bit. 64-bit is desired, since each CTM element is
S31.32 format, but there's no xserver support for it.

We can use the first 32 bits to represent the S31 signed-
magnitude whole component, and the second 32 bits to represent
the fractional component.

See the comments in set_ctm() for details.

Signed-off-by: Leo (Sunpeng) Li <sunpeng.li@amd.com>